% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent_summary.R
\name{sent_summary}
\alias{sent_summary}
\title{Sentiment Summariser}
\usage{
sent_summary(sentiment = "positive", df = tweetsDF, n = 3,
  meth = "afinn")
}
\arguments{
\item{sentiment}{'positive' or 'negative' currently working}

\item{df}{dataframe containing the tweets}

\item{n}{number of tweets to summarise}

\item{order.desc}{TRUE is default which outputs highest scores i.e. most positive tweets. FALSE
outputs the most negative tweets}
}
\description{
This function summarises the sentiment calculations and outputs the top scores
This can be ordered as positive tweets first or negative tweets first (ascending or descending)
}
\examples{
sent.summary('positive', n = 6)
}
\keyword{sentiment}
\keyword{summary}

