% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_clean.R
\name{text_clean}
\alias{text_clean}
\title{Tweet Cleaner}
\usage{
text_clean(x = tweetDF$text)
}
\arguments{
\item{x}{A vector of tweets}
}
\description{
This function cleans the raw tweets scraped from twitter, to suit the hash_sentiment function.
}
\examples{
text.clean(x = tweetDF$text)
}
\keyword{clean}
\keyword{tweet,}
\keyword{twitter,}

