% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizers.R
\name{token.nrc}
\alias{token.nrc}
\title{Word Scorer}
\usage{
token.nrc(tweet.vec = x)
}
\arguments{
\item{tweet.vec}{vector of tweets as strings}
}
\description{
These two functions score each word in each tweet and returns a list of dataframes - one dataframe for each
tweet
}
\examples{
token.nrc(tweet.vec = tweetsDF$text)


}
\keyword{tokenizer}

